+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1-dev
 
 - Fix: root component was not added to ancestor list.
+- Enable multiple (and non-component) children as root for a `View`.
 
 ## 0.1.0